+++ test/dygraph_to_static/test_train_step_resnet18_adam.py
@@ -15,7 +15,7 @@
 import platform
 import unittest
 
-from test_train_stepx import (
+from test_train_step import (
     TestTrainStepTinyModel,
     loss_fn_tiny_model,
     train_step_tiny_model,
@@ -25,12 +25,12 @@ import paddle
 from paddle.vision.models import resnet18
 
 
-class TestTrainStepResNet18Sgd(TestTrainStepTinyModel):
+class TestTrainStepResNet18Adam(TestTrainStepTinyModel):
     def setUp(self):
         self.input = paddle.randn([64, 3, 224, 224])
         self.net_creator = resnet18
         self.lr_creator = lambda: 0.001
-        self.optimizer_creator = paddle.optimizer.SGD
+        self.optimizer_creator = paddle.optimizer.Adam
         self.loss_fn = loss_fn_tiny_model
         self.train_step_func = train_step_tiny_model
         self.steps = 3
